Refactor File Structure (PODs)
